+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
+### Changed
+
+- Improve Accordion component keeps toggle in view when toggling 
+
 ### Fixed
 
 - Fix Slider drag stops on scroll now
-- Fix Accordion scrolls toggle into view if needed
 
 ## 3.15.6 (August 31, 2022)